+++ Common/src/main/java/fuzs/puzzleslib/client/renderer/DynamicBuiltinModelItemRenderer.java
@@ -6,13 +6,13 @@ import net.minecraft.client.renderer.block.model.ItemTransforms;
 import net.minecraft.world.item.ItemStack;
 
 /**
- * {@link net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer} as a functional interface
+ * {@link net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer#renderByItem} as a functional interface
  */
 @FunctionalInterface
 public interface DynamicBuiltinModelItemRenderer {
 
     /**
-     * Renders an item stack, kindly copied from Fabric Api :)
+     * renders an item stack, kindly copied from Fabric Api :)
      *
      * @param stack             the rendered item stack
      * @param mode              the model transformation mode
@@ -21,5 +21,5 @@ public interface DynamicBuiltinModelItemRenderer {
      * @param light             packed lightmap coordinates
      * @param overlay           the overlay UV passed to {@link com.mojang.blaze3d.vertex.VertexConsumer#overlayCoords}
      */
-    void render(ItemStack stack, ItemTransforms.TransformType mode, PoseStack matrices, MultiBufferSource vertexConsumers, int light, int overlay);
+    void renderByItem(ItemStack stack, ItemTransforms.TransformType mode, PoseStack matrices, MultiBufferSource vertexConsumers, int light, int overlay);
 }