+++ Common/src/main/java/fuzs/puzzleslib/util/PuzzlesUtil.java
@@ -5,6 +5,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Collection;
+import java.util.ServiceLoader;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
@@ -15,9 +16,12 @@ import java.util.function.Supplier;
 public class PuzzlesUtil {
 
     /**
-     * @param object object to consume
-     * @param consumer action
-     * @param <T> object type
+     * perform <code>action</code> on a (most likely newly created) <code>object</code>, then return it
+     *
+     * @param object        object to consume
+     * @param consumer      action
+     * @param <T>           object type
+     *
      * @return the object
      */
     public static <T> T make(T object, Consumer<T> consumer) {
@@ -27,10 +31,12 @@ public class PuzzlesUtil {
 
     /**
      * run an action, if an exception occurs run a different action
-     * @param object object for actions
-     * @param action action to attempt
-     * @param orElse action in case <code>action</code> throws an exception
-     * @param <T> type of object
+     *
+     * @param object        object for actions
+     * @param action        action to attempt
+     * @param orElse        action in case <code>action</code> throws an exception
+     * @param <T>           type of object
+     *
      * @return was there an exception
      */
     public static <T> boolean runOrElse(@NotNull T object, Consumer<T> action, Consumer<T> orElse) {
@@ -46,9 +52,11 @@ public class PuzzlesUtil {
 
     /**
      * perform action for nullable object
-     * @param object nullable object
-     * @param action action to perform
-     * @param <T> type of object
+     *
+     * @param object        nullable object
+     * @param action        action to perform
+     * @param <T>           type of object
+     *
      * @return was <code>object</code> null
      */
     public static <T> boolean acceptIfPresent(@Nullable T object, Consumer<T> action) {
@@ -61,10 +69,12 @@ public class PuzzlesUtil {
 
     /**
      * get an instance that might not have been created yet
-     * @param instance instance to get, might have to be created
-     * @param supplier supplier for creating instance
-     * @param consumer save created instance
-     * @param <T> instance type
+     *
+     * @param instance      instance to get, might have to be created
+     * @param supplier      supplier for creating instance
+     * @param consumer      save created instance
+     * @param <T>           instance type
+     *
      * @return the instance
      */
     public static <T> T getOrElse(@Nullable T instance, Supplier<T> supplier, Consumer<T> consumer) {
@@ -77,9 +87,11 @@ public class PuzzlesUtil {
 
     /**
      * might want to wrap this in an optional
-     * @param collection collection to get entry from
-     * @param weight weight getter function
-     * @param <T> type of objects in collection
+     *
+     * @param collection        collection to get entry from
+     * @param weight            weight getter function
+     * @param <T>               type of objects in collection
+     *
      * @return random entry or null when collection is empty
      */
     @Nullable
@@ -97,12 +109,29 @@ public class PuzzlesUtil {
     }
 
     /**
-     * @param toRound number to round
-     * @param decimalPlaces amount of decimal places
+     * rounds a number duh
+     *
+     * @param toRound           number to round
+     * @param decimalPlaces     amount of decimal places
+     *
      * @return rounded number
      */
     public static double round(double toRound, int decimalPlaces) {
         final double power = Math.pow(10, decimalPlaces);
         return  Math.round(toRound * power) / power;
+    }
+
+    /**
+     * loads a service provider interface yay
+     *
+     * @param clazz         service provider interface class to load
+     * @param <T>           interface type
+     *
+     * @return loaded service
+     */
+    public static <T> T loadServiceProvider(Class<T> clazz) {
+        return ServiceLoader.load(clazz)
+                .findFirst()
+                .orElseThrow(() -> new NullPointerException("Failed to load service for " + clazz.getName()));
     }
 }