+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.2.13-1.19.2] - 2022-08-29
+### Added
+- Allow for registering listeners that run after baked models have been reloaded
+- This includes a helper method for baking a model in the common project
+### Changed
+- Mod service classes should no longer extend `CoreServices`, method for loading a service is now accessible from outside
+### Removed
+- Removed `ModelEvents` class on Fabric, the same can be achieved by registering a `ReloadListener`
+
 ## [v4.2.12-1.19.2] - 2022-08-28
 ### Added
 - Allow for registering item decorators (like stack count) in common project, includes api implementation on Fabric